Run cargo fmt
